third_party/WebKit/LayoutTests/external/wpt/html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-tokenization-noopener-expected.txt

This is a testharness.js-based test.
FAIL tokenization should skip window features separators before `name`assert_equals: "noopener" should activate feature "noopener" expected null but got object "[object Window]"
PASS feature `name` should be converted to ASCII lowercase
FAIL after `name`, tokenization should skip window features separators that are not "=" or ","assert_equals: "foo noopener=1" should activate feature "noopener" expected null but got object "[object Window]"
PASS Tokenizing should ignore window feature separators except "," after initial "=" and before value
PASS Tokenizing should read characters until first window feature separator as `value`
PASS "noopener" should be based on name (key), not value
FAIL invalid feature names should not tokenize as "noopener"assert_not_equals: "\0noopener" should NOT activate feature "noopener" got disallowed value null
Harness: the test ran to completion.
